fix conditional resource referencing error
According to azure document,
"If you use the reference function in a resource that is conditionally
deployed, the function is evaluated even if the resource isn't deployed.
You get an error if the reference function refers to a resource that doesn't
exist. Use the if function to make sure the function is only evaluated when
the resource is being deployed."