Merge HA changes back to master (#14)
* Squash merge in HA-LoadBalancer changes

* Update HA-LoadBalancer template URL in readme

* fix conditional resource referencing error

According to azure document,
"If you use the reference function in a resource that is conditionally
deployed, the function is evaluated even if the resource isn't deployed.
You get an error if the reference function refers to a resource that doesn't
exist. Use the if function to make sure the function is only evaluated when
the resource is being deployed."

* Keep only Gen7 support

Gen6 support is in a separate template.

* Revert "Update HA-LoadBalancer template URL in readme"

This reverts commit 03db9bc2aeeaa21285bba6e0fe525d994fc9df25.

* HA: map 270 to D3 or DS3 machine size