HA: fix InvalidTemplate issue when existing vnet is used
The deployment may fail with error property 'addressPrefix' doesn't exist.
The existing vnet may have been created with property "addressPrefixes"
instead of "addressPrefix".
The object returned by reference(variables('lanSubnetID'), '2018-04-01')
does not contain the addressPrefixes key, because the API version is older
than the introduction of addressPrefixes.
Upgrade the API version so the addressPrefixes can be read. Also use
tryGet() to read the properties. Because addressPrefix and addressPrefixes
cannot coexist, one tryGet() will return null and the other will return the
expected property value. When either of the properties is assigned null, it'll
be ignored, while the other will be assigned the expected value.