HA: Fix default public IP DNS name DnsRecordInUse error
If an existing Resource Group is used and there is a leftover public
IP in the RG, creating a new public IP in that RG will cause
DnsRecordInUse error. This is because the DNS name is constructed
from a unique hash string generated using only the RG ID. Two public
IPs in the same RG will have the same unique hash string. The fix is
to use RG ID as well as public IP resource name to generate the
unique string.